Rename tooltip to custom-tooltip
